US116101
Reorganized file layout so that ONLY the "./public" directory is served out by Node.JS.
added ./ssl directory to store the certificate and private key to enable SSL.  This is now protected due to the folder re-organization.
Configured server.js to host an HTTPS server
Created a .gitignore to ignore the SSL directory.  We do not want the public key in the GitHub repository.
placed .gitignore under version control
Updated main.js with the ICE URI for the TURN/STUN service

Configured TURN service for TLS/DTLS